Fix global config load for windows, move load file to util add unit tests to load file method
