+++ fixtures/fixture.javascript.js
@@ -1,0 +1,3 @@
+module.exports = {
+  data: 1,
+}
